oracle problems workaround

+++ CondFormats/HcalObjects/interface/HcalZSThreshold.h
@@ -13,12 +13,16 @@ class HcalZSThreshold
  HcalZSThreshold(): mId(0), mLevel(0), empty(true) {}
  HcalZSThreshold(unsigned long fid, int level): mId(fid), mLevel(level), empty(false) {}
 
-  uint32_t rawId() const {return mId;}
+    // because of an oracle digestion problem with uint32_t 
+    // use unsigned long long
+  //  uint32_t rawId() const {return mId;}
+  unsigned long long rawId() const {return mId;}
   int getValue() const {return mLevel;}
   bool isEmpty() const {return empty;}
 
  private:
-  uint32_t mId;
+  //  uint32_t mId;
+  unsigned long long mId;
   int mLevel;
   bool empty;
 };
